Enable using byebug w/ Docker
This allows a connection to the running container when using byebug. If
the current web app container is attr_json_test_web_1  then when a
byebug breakpoint is hit, open a terminal/tab and run:

docker attach attr_json_test_web_1

This will give you a prompt for using byebug. Hooray